添加全局异常处理器，进一步实现员工的管理，完善RBAC模型的多角色部分,进一步配置spring security

+++ src/main/java/com/easyauth/service/UserService.java
@@ -6,5 +6,5 @@ import com.easyauth.domain.entity.User;
 public interface UserService extends IService<User> {
     public User getByUsername(String username);
 
-    void switchStatus(Long id, Integer status);
+    void switchStatus(Long id, Long status);
 }
